// src/generated/Command.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Command = { "type": "triggerCue", cueId: string, } | { "type": "stopPlayback", playbackId: string, } | { "type": "fadePlayback", playbackId: string, } | { "type": "stopAll" } | { "type": "fadeAll", durationMs: number, } | { "type": "setMasterVolume", volume: number, };